+++ omeroweb/webadmin/views.py
@@ -38,7 +38,7 @@ import omeroweb.webclient.views
 from omeroweb.version import omeroweb_buildyear as build_year
 from omeroweb.version import omeroweb_version as omero_version
 
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 from django.views.decorators.debug import sensitive_post_parameters
 from django.http import HttpResponseRedirect
 from django.utils.translation import ugettext as _
@@ -256,18 +256,28 @@ def drivespace_json(
             "from Pixels p join p.pixelsType as pt join p.image i "
             "left outer join i.fileset f "
             "join p.details.owner as owner "
+            "join p.details.group as group "
             "where f is null"
         )
 
         filesQuery = (
             "select sum(origFile.size) from OriginalFile as origFile "
+            "join origFile.details.group as group "
             "join origFile.details.owner as owner"
         )
 
+        clauses = []
+        group = ctx.getOmeroGroup()
+        # the group context won't filter out files in 'user' group
+        # so we explicitly filter by group ID
+        if group != "-1":
+            params.add("gid", omero.rtypes.rlong(group))
+            clauses.append("group.id = (:gid)")
         if eid is not None:
             params.add("eid", omero.rtypes.rlong(eid))
-            pixelsQuery = pixelsQuery + " and owner.id = (:eid)"
-            filesQuery = filesQuery + " where owner.id = (:eid)"
+            clauses.append("owner.id = (:eid)")
+        pixelsQuery = pixelsQuery + " and " + (" and ".join(clauses))
+        filesQuery = filesQuery + " where " + (" and ".join(clauses))
         # Calculate disk usage via Pixels
         result = queryService.projection(pixelsQuery, params, ctx)
         if len(result) > 0 and len(result[0]) > 0:
@@ -317,11 +327,26 @@ def drivespace_json(
     # users within a single group
     elif groupId is not None:
         ctx.setOmeroGroup(groupId)
-        for e in conn.getObjects("Experimenter"):
+        exps_in_group = [
+            e.id
+            for e in conn.getObjects(
+                "Experimenter",
+                opts={"load_experimentergroups": False, "experimentergroup": groupId},
+            )
+        ]
+        for e in conn.getObjects(
+            "Experimenter", opts={"load_experimentergroups": False}
+        ):
             b = getBytes(ctx, e.getId())
+            memberOfGroup = e.getId() in exps_in_group
             if b > 0:
                 diskUsage.append(
-                    {"label": e.getNameWithInitial(), "data": b, "userId": e.getId()}
+                    {
+                        "label": e.getNameWithInitial(),
+                        "data": b,
+                        "userId": e.getId(),
+                        "memberOfGroup": memberOfGroup,
+                    }
                 )
 
     diskUsage.sort(key=lambda x: x["data"], reverse=True)