+++ omeroweb/webclient/tree.py
@@ -537,10 +537,11 @@ def _marshal_image(
       * details.permissions (dict)
       * fileset_id (rlong)
 
-    May also take a row_pixels (list) if X,Y,Z dimensions are loaded
+    May also take a row_pixels (list) if X,Y,Z,T dimensions are loaded
       * pixels.sizeX (rlong)
       * pixels.sizeY (rlong)
       * pixels.sizeZ (rlong)
+      * pixels.sizeT (rlong)
 
     @param conn OMERO gateway.
     @type conn L{omero.gateway.BlitzGateway}
@@ -559,10 +560,11 @@ def _marshal_image(
     if fileset_id_val is not None:
         image["filesetId"] = fileset_id_val
     if row_pixels:
-        sizeX, sizeY, sizeZ = row_pixels
+        sizeX, sizeY, sizeZ, sizeT = row_pixels
         image["sizeX"] = unwrap(sizeX)
         image["sizeY"] = unwrap(sizeY)
         image["sizeZ"] = unwrap(sizeZ)
+        image["sizeT"] = unwrap(sizeT)
     if share_id is not None:
         image["shareId"] = share_id
     if date is not None:
@@ -657,6 +659,7 @@ def marshal_images(
              ,
              pix.sizeX as sizeX,
              pix.sizeY as sizeY,
+             pix.sizeT as sizeT,
              pix.sizeZ as sizeZ
              """
 
@@ -759,7 +762,7 @@ def marshal_images(
         ]
         kwargs = {"conn": conn, "row": d[0:5]}
         if load_pixels:
-            d = [e["sizeX"], e["sizeY"], e["sizeZ"]]
+            d = [e["sizeX"], e["sizeY"], e["sizeZ"], e["sizeT"]]
             kwargs["row_pixels"] = d
         if date:
             kwargs["acqDate"] = e["acqDate"]
@@ -1502,6 +1505,7 @@ def marshal_tagged(
              ,
              pix.sizeX as sizeX,
              pix.sizeY as sizeY,
+             pix.sizeT as sizeT,
              pix.sizeZ as sizeZ
              """
         extraObjs = " left outer join obj.pixels pix"
@@ -1540,7 +1544,7 @@ def marshal_tagged(
         ]
         kwargs = {}
         if load_pixels:
-            d = [e[0]["sizeX"], e[0]["sizeY"], e[0]["sizeZ"]]
+            d = [e[0]["sizeX"], e[0]["sizeY"], e[0]["sizeZ"], e[0]["sizeT"]]
             kwargs["row_pixels"] = d
         if date:
             kwargs["acqDate"] = e[0]["acqDate"]